+++ xbacnet-server/settings.py
@@ -1,0 +1,10 @@
+xbacnet = {
+    'user': 'root',
+    'password': '!XBACnet1',
+    'host': '127.0.0.1',
+    'database': 'xbacnet',
+}
+
+# interval for object persistence task
+PERSISTENCE_INTERVAL = 5.0
+REFRESHING_INTERVAL = 5.0