+++ CHANGELOG.md
@@ -11,8 +11,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [v1.0.0] -   2024-12-08
 ### Added
--   added database sql script.
-
+- added database sql script.
+- added xbacnet server
 ### Changed
 -   None.